refactor(max-zoom): rename max-zoom hook to max-zoom-service
Breaking Changes
