+++ library/src/hooks/max-zoom-service.ts
